### 2024-01-29 #6302 @editorialbot
[PRE REVIEW]: agweather-qaqc: An Interactive Python Package for Quality Assurance and Quality Control of Daily Agricultural Weather Data and Calculation of Reference Evapotranspiration
**Submitting author:** @cwdunkerly  (<a href="http://orcid.org/0000-0003-3592-4118">Christian Dunkerly</a>)
**Repository:** https://github.com/WSWUP/agweather-qaqc
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.2

Thanks for submitting your paper to JOSS @cwdunkerly . **Currently, there isn't a JOSS editor assigned** to your paper.
