### 2024-02-18 #6368 @editorialbot
[REVIEW]: agweather-qaqc: An Interactive Python Package for Quality Assurance and Quality Control of Daily Agricultural Weather Data and Calculation of Reference Evapotranspiration
**Submitting author:** @cwdunkerly  (<a href="http://orcid.org/0000-0003-3592-4118">Christian Dunkerly</a>)
**Repository:** https://github.com/WSWUP/agweather-qaqc
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.4
