### 2024-02-18 WSWUP/agweather-qaqc#18 @dlebauer
Add high level description of docs to readme  
I’m conducting review for JOSS and will add comments as issues as I go. https://github.com/openjournals/joss-reviews/issues/6368

README should contain a high level overview of the documentation. See https://joss.readthedocs.io/en/latest/review_criteria.html for details
