### 2024-02-26 WSWUP/agweather-qaqc#19 @dostuffthatmatters
Corrections in Paper Text
_This issue is related to the review process for JOSS: https://github.com/openjournals/joss-reviews/issues/6368_

Below, you can find some things in the paper that could be corrected. Some of them are rather stylistic:

- [x] Higher resolution image files for Figures 1 and 2; maybe as SVG/EPS?
- [x] Lines 99-100: _"While ease-of-use for a non-technical user was one of the principal goals, the software workflow can be automated with the inclusion of libraries such as PyAutoGUI."_ Automated as in checking for trends/anomalies visually in an automated way?
- [x] Lines 102-109, 116-119, and 125-128 are a repetition of lines 56-98. You could introduce the bullet point list as "In the already described process, the package also performs the following steps: ..." instead of an "all features" list. Since 80-81 does not entirely cover 125-128, you could expand the sentence in 80-81.
- [x] Figure 2: I would prefer having some plot titles (more descriptive than "A", "B", and "C").
- [x] I would move the sections "State of the Field" and "Research enabled by agweather-qaqc" after the "Statement of Need".
- [x] The paper would benefit from a grammar check. I didn't find many, but there are some grammatical errors.

_ADDITION:_ From the repository, it is obvious, that you are the main developer and maintainer of the package. Could you dedicate a small section in the paper to mentioning the roles of the co-authors of the paper?

- [x] clarify the roles of the different authors
