### 2024-03-08 WSWUP/agweather-qaqc#20 @dlebauer
Add contributor guidelines
Regarding JOSS review https://github.com/openjournals/joss-reviews/issues/6368

JOSS requires

> Clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Consider adding CONTRIBUTING.md and CODE_OF_CONDUCT files as well as brief information about contributing to the README

See https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project and https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
