### 2024-03-08 f5cb79e
updates to installation instructions
