### 2024-03-08 WSWUP/agweather-qaqc#20 @dostuffthatmatters
@cwdunkerly

You might find issue templates helpful to provide a few clear methods to communicate bugs/ideas.

On the index page of the docs you mention "agweather-qaqc was developed for daily weather data. Support for sub-daily data may be added eventually." you can add links to the issues page to comments like this.

I personally like a public backlog of ideas. Especially on research software that lacks funds to implement feature for others, you could label issues as "contributions welcome".
