### 2024-03-08 WSWUP/agweather-qaqc#24 @dostuffthatmatters
Test Setup Review
_This issue is related to the review process for JOSS: https://github.com/openjournals/joss-reviews/issues/6368_

You already have a suite of PyTest test cases which is very good. All tests passed on my system on the first try.

- [x] Use a more established directory structure

I would recommend moving the file `test_calculations.py` and the directory `test_files` into a `tests` directory. Maybe also split up the test file into multiple test files - this is quite stylistic though. This repo follows a widely used project structure: https://github.com/navdeep-G/samplemod

```
README.md
...
tests/
    test_files/
    test_io.py
    test_conversions.py
    test_calculations.py
```

- [x] Add pytest to CI

You should run the tests using GitHub actions. If you choose to use PDM, you can simply use Pip to install the dependencies inside the CI and don't have to bother with setting up Conda/PDM/Pipenv/Poetry.

- [x] Run tests for different python versions (matrix)

The GitHub CI supports running tests for multiple Python versions in parallel

- [x] Run the end to end test on your test data in the CI

I.e. running `qaqc_single_station.py` using the default config. You would have to add an option to skip/do the corrections without a user input via keyboard. You can pipe a stdin to your script using `echo 0 | python qaqc_single_station.py`.
