### 2024-03-20 #6368 @cwdunkerly
@cheginit I have been working on this slowly, however I've been sitting on some of the changes without pushing them.

Let me look over everything and get them posted. Both @dlebauer and @dostuffthatmatters 's comments have been very clear and helpful.
