### 2024-03-21 3fd3f5f
added alternative installation methods, added support for 3.12+
Will add installs to documentation soon
