+++ qaqc_single_station.py
@@ -8,17 +8,12 @@ if __name__ == "__main__":
     # or it can serve as an example of how to make calls to the AgWeatherQAQC class
 
     # Check if python version is acceptable
-    if sys.version_info.major < 3 or (sys.version_info.major == 3 and sys.version_info.minor < 9):
-        raise SystemError(
-            f'\n\nagweatherqaqc requires a python version of 3.9.17 or newer. \n'
-            f'The current version of python being run is {sys.version}. \n\n')
-    elif sys.version_info.major > 3 or (sys.version_info.major == 3 and sys.version_info.minor > 11):
-        print(
-            f'\n\nag-weather-qaqc has been tested on python versions 3.9.17 through 3.11.X. \n' 
-            f'The current version of python being run is {sys.version}. \n\n'
-            f'The script will likely still function, but spend extra time verifying outputs.')
-    else:
+    if sys.version_info.major == 3 and sys.version_info.minor >= 9:
         pass
+    else:
+        raise SystemError(
+            f'\n\nagweatherqaqc requires a python version between 3.9.X and 3.X.X. \n'
+            f'The current version of python being run is {sys.version}. \n\n')
 
     # Check if user has passed in a config file, or else just grab the default.
     # Also see if user has passed a metadata file to allow for automatic reading/writing into the metadata file.
