+++ requirements.txt
@@ -1,0 +1,10 @@
+bokeh >= 3.0
+numpy >= 1.2
+openpyxl >= 3.0
+pandas >= 1.5
+pathlib >= 1.0
+pytest >= 7.0
+refet >= 0.4.0
+xlsxwriter >= 3.0
+sphinx
+sphinx_rtd_theme
