### 2024-03-21 WSWUP/agweather-qaqc#22 @cwdunkerly
When you have a moment, please pull the latest and check the new .gitignore against your venv files.

I used the toptal tool for Windows, Python, MacOS, and Linux, as well as any specific files I knew would be created by running the code.
