### 2024-03-21 f0237d3
Updated readme with description of software
