### 2024-03-22 WSWUP/agweather-qaqc#18 @cwdunkerly
Sorry I've let this one sit so long, I believe the new version of the readme is sufficient, however if you disagree, please reopen the issue.
