+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,14 @@
+---
+name: Feature request
+about: Suggest an idea
+title: "[Feature Request]"
+labels: enhancement
+assignees: ''
+
+---
+
+**Please describe the feature you'd like to see**
+A clear and concise description of what the problem is and how you'd like to see it fixed.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
