+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,0 +1,23 @@
+---
+name: Bug Report
+about: Found a bug, issue, or error.
+title: "[BUG]"
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**Data Sample**
+If your issue is with processing some data, please include a sample (the first~10 lines) of your data file here.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Python Version Info**
+Please include a list of the version numbers for Python and any relevant libraries, as well as your OS.
+
+**Additional context**
+Please add any other context about the problem here.
