### 2024-03-26 27d4c92
Create CONTRIBUTING.md
