+++ .github/CONTRIBUTING.md
@@ -1,0 +1,26 @@
+## Contributing to ``agweather-qaqc``
+
+``agweather-qaqc`` is open-source, contributions are both welcome and encouraged!
+
+#### **Finding A Bug**
+
+* **Please check to see if the [issue](https://github.com/WSWUP/agweather-qaqc/issues) already exists.**
+
+* If the bug is new, please open a [new issue](https://github.com/WSWUP/agweather-qaqc/issues/new/choose), and provide as much relevant information as possible. If your issue has to do with the data you're processing, please include the first few lines of the file.
+
+#### **Fixing/Patching A Bug**
+
+* Please open a new GitHub pull request, and include a clear description of the issue and the solution.
+
+* ``agweather-qaqc`` makes an effort to follow the [PEP 8](https://peps.python.org/pep-0008/) style guide. Pull requests that significantly deviate from this style may not be accepted.
+
+* Pull requests for cosmetic fixes may not be accepted, depending on the scope of the changes.
+
+#### **Proposing/Requesting a New Feature**
+
+* **Please check to see if the [feature request](https://github.com/WSWUP/agweather-qaqc/issues) already exists.**
+
+* Open a [new request](https://github.com/WSWUP/agweather-qaqc/issues/new/choose) and describe the issue. Please do this **even if you plan to code the feature yourself** to prevent possible duplication of effort.
+  
+#### Questions / Comments / Feedback
+* Please email me at Christian.Dunkerly@dri.edu.
