### 2024-03-26 WSWUP/agweather-qaqc#20 @cwdunkerly
I have added a CONTRIBUTING.md in .github/ .

I have also created two issue templates, one for problems/bugs, one for suggested features. I have also created a new issue describing a feature request for sub-daily data.

@dlebauer I have looked over the example CODE_OF_CONDUCT files and I feel as though this may be out of the scope for this repository. (Not that I disagree with the intent behind such a document)

I will close this issue as completed on Wednesday morning, unless either of you feel strongly otherwise regarding the code of conduct.
