### 2024-03-27 #6368 @cwdunkerly
Just reporting in, I am currently working on [issue 24](https://github.com/WSWUP/agweather-qaqc/issues/24), and I have set aside this weekend to complete all outstanding issues that I don't get to during this work week.
