### 2024-03-29 WSWUP/agweather-qaqc#20 @dlebauer
Hi @cwdunkerly code of conduct isn't required, its up to you. But I'll note that the template isn't as useful as the docs https://opensource.guide/code-of-conduct/#why-do-i-need-a-code-of-conduct.
