### 2024-04-01 83ff5a1
Changed location of test files
Also changed default behavior of qaqc_single_station.py, will now read in the config in test if none is provided

Edited .gitignore to allow updated paths for test files

Removed todos that had already been fixed in test_calculations.py, will track future implementation plans as issues instead
