+++ tests/test_files/test_config.ini
@@ -5,7 +5,7 @@
 
 
 # SPECIFY THE PATH TO THE DATA FILE (MUST BE .csv, .xls, OR .xlsx)
-DATA_FILE_PATH = test_files/test_data.csv
+DATA_FILE_PATH = tests/test_files/test_data.csv
 
 
 # STATION LATITUDE AND LONGITUDE (IN DECIMAL DEGREES)
