### 2024-04-01 51e14b1
added initial CI file, updated paths in test_calculations.py
