+++ .github/workflows/ci.yml
@@ -1,0 +1,38 @@
+name: CI
+
+on:
+  push:
+  pull_request:
+
+jobs:
+  run-tests:
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version:
+          - "3.9"
+          - "3.10"
+          - "3.11"
+          - "3.12"
+
+    name: run_tests
+    runs-on: ${{ matrix.os }}
+
+    steps:
+      - name: checkout_code
+        uses: actions/checkout/@v4
+
+      - name: setup_python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python_version }}
+
+      - name: install_dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install disutils
+          pip install -r requirements.txt
+
+      - name: run_pytest
+        run: python -m pytest
