### 2024-04-01 7df96db
added additional temporary test to test_calculations.py to experiment with CI versioning
