### 2024-04-01 69b0efc
removed temporary test from last commit, added calling qaqc_single_station.py to CI
