+++ .github/workflows/ci.yml
@@ -36,3 +36,6 @@ jobs:
 
       - name: run_pytest
         run: python -m pytest
+
+      - name: run_qaqc_single_station
+        run: echo 0 | python qaqc_single_station.py
