### 2024-04-01 953b0f1
reconfigured and renamed workflows
Running qaqc_single_station.py across the full matrix was consuming too much action time
