### 2024-04-01 WSWUP/agweather-qaqc#24 @cwdunkerly
Finished all tasks, I split the end-to-end test and the pytest into separate workflows, testing the end-to-end across the full matrix of operating systems was consuming a lot of action time.
