### 2024-04-01 c19764f
Updated CONTRIBUTING.md to include a code of conduct
