+++ .github/CONTRIBUTING.md
@@ -2,6 +2,10 @@
 
 ``agweather-qaqc`` is open-source, contributions are both welcome and encouraged!
 
+#### **Code of Conduct**
+Please act in a welcoming, professional manner when contributing to this project, whether you are documenting new issues, contributing to discussion, or making pull requests.  Please ensure that feedback is constructive and be respectful of differing viewpoints. You can report any instances of unacceptable behavior by emailing Christian.Dunkerly@dri.edu. All issues will be investigated promptly.
+**Violations of this code of conduct will result in a response ranging from a warning to a ban, depending on the severity and/or frequency.**
+
 #### **Finding A Bug**
 
 * **Please check to see if the [issue](https://github.com/WSWUP/agweather-qaqc/issues) already exists.**
@@ -23,4 +27,4 @@
 * Open a [new request](https://github.com/WSWUP/agweather-qaqc/issues/new/choose) and describe the issue. Please do this **even if you plan to code the feature yourself** to prevent possible duplication of effort.
   
 #### Questions / Comments / Feedback
-* Please email me at Christian.Dunkerly@dri.edu.
+* Please send an email to Christian.Dunkerly@dri.edu.
