### 2024-04-01 WSWUP/agweather-qaqc#20 @cwdunkerly
I understand, that link was helpful. I have written a smaller paragraph on a code of conduct and included it in the contributing.md file.
