### 2024-04-01 WSWUP/agweather-qaqc#22 @cwdunkerly
I have .csv's as an option for output files, and I have created an issue for future development over implementing static images as outputs.

The latter will require more work as the HTML version is used as part of identifying intervals of data that need to be corrected.
