### 2024-04-01 WSWUP/agweather-qaqc#24 @dostuffthatmatters
> Finished all tasks, I split the end-to-end test and the pytest into separate workflows, testing the end-to-end across the full matrix of operating systems was consuming a lot of action time.

Great!

I think it is sufficient it you test the package on `ubuntu-latest`. Linux CI workers typically spin up way faster than MacOS or Windows workers. You can cache pip dependencies, and - if applicable - also use the `pyproject.toml` file as a dependency list for pip.

Below you can find a few changes I suggest for your `.github/workflows/ci_pytest.yml`. I included the test step for `qaqc_single_station.py` because then you can save some CI time in installation.

```diff
jobs:
  run-tests:
    strategy:
      fail-fast: false
      matrix:
-       os: [ubuntu-latest, macos-latest, windows-latest]
        python-version:
          - "3.9"
          - "3.10"
          - "3.11"
          - "3.12"

    name: run_tests
-   runs-on: ${{ matrix.os }}
+   runs-on: ubuntu-latest

    steps:
      - name: checkout_code
        uses: actions/checkout/@v4

      - name: setup_python
        uses: actions/setup-python@v5
        with:
          python-version: ${{ matrix.python_version }}
+         cache: 'pip'
+         cache-dependency-path: 'pyproject.toml'
      - name: install_dependencies
        run: |
          pip install -r pyproject.toml

      - name: run_pytest
        run: python -m pytest

+     - name: run_qaqc_single_station
+       run: echo 0 | python qaqc_single_station.py
```
