### 2024-04-01 #6368 @cwdunkerly
The last issue outstanding is [#19](https://github.com/WSWUP/agweather-qaqc/issues/19) which pertains to paper text corrections.
I'm resolving this tomorrow, I'm just conferring with the other authors first.

Once this is done, what is the next step? The only other issue I am aware of is adding the other installation methods (pipenv, PDM) to the docs on the github.io page, which I will also have resolved tomorrow.

If I have missed issues detailed somewhere else I apologize.
