### 2024-04-01 #6368 @dostuffthatmatters
I only have to parse the bibliography and check for any non-cited sources/comparable packages. I will try to get that done by tomorrow.
