### 2024-04-02 8b6e382
added option for saving output as .csv file in case user prefers it over excel, updated CI process
Updated CI process to optimize action usage time, thank you to @dostuffthatmatters for the suggestion
