+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: ci_pytest
+name: ci_tests
 
 on:
   push:
@@ -9,7 +9,6 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version:
           - "3.9"
           - "3.10"
@@ -17,7 +16,7 @@ jobs:
           - "3.12"
 
     name: run_tests
-    runs-on: ${{ matrix.os }}
+    runs-on: ubuntu-latest
 
     steps:
       - name: checkout_code
@@ -27,11 +26,14 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python_version }}
+          cache: 'pip'
+          cache-dependency-path: 'pyproject.toml'
 
       - name: install_dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
+        run: pip install -r pyproject.toml
 
       - name: run_pytest
         run: python -m pytest
+
+      - name: run_qaqc_single_station
+        run: echo 0 | python qaqc_single_station.py
