+++ tests/test_files/test_config.ini
@@ -33,7 +33,6 @@ MISSING_OUTPUT_VALUE = nan
 LINES_OF_HEADER = 1
 LINES_OF_FOOTER = 0
 
-
 ############################################################################################################################
 ############################################################################################################################
 [OPTIONS]
@@ -52,6 +51,10 @@ AUTOMATIC_OPTION = 0
 #	0 - DO NOT FILL MISSING DATA
 #	1 - FILL MISSING DATA
 FILL_OPTION = 0
+
+
+# OUTPUT FILE FORMAT FOR CORRECTED DATA - MUST BE EITHER 'CSV' OR 'XLSX'
+OUTPUT_DATA_FORMAT = XLSX
 
 
 ############################################################################################################################
