+++ .github/workflows/ci_run_qaqc_single_station.yml
@@ -1,37 +1,0 @@
-name: ci_run_qaqc_single_station
-
-on:
-  push:
-  pull_request:
-
-jobs:
-  run-tests:
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [windows-latest]
-        python-version:
-          - "3.9"
-          - "3.10"
-          - "3.11"
-          - "3.12"
-
-    name: run_tests
-    runs-on: ${{ matrix.os }}
-
-    steps:
-      - name: checkout_code
-        uses: actions/checkout/@v4
-
-      - name: setup_python
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python_version }}
-
-      - name: install_dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-
-      - name: run_qaqc_single_station
-        run: echo 0 | python qaqc_single_station.py
