### 2024-04-02 13cb54b
attempt at fixing missing section of pyproject.toml
