+++ pyproject.toml
@@ -1,6 +1,10 @@
+[build-system]
+requires = ["pdm-backend"]
+build-backend = "pdm.backend"
+
 [project]
 name = "agweatherqaqc"
-version = "1.0.3"
+version = "1.0.4"
 description = "Required Packages for the agweatherqaqc library"
 authors = [
     {name = "Christian Dunkerly", email = "christian.dunkerly@dri.edu"},
