### 2024-04-02 ab6d339
second attempt at fixing CI
