### 2024-04-02 WSWUP/agweather-qaqc#24 @cwdunkerly
@dostuffthatmatters  That makes sense, thank you for the suggestion. I do have one question:

I understand the advantages of PDM/pyproject.toml, however when I used your setup as written I got the following errors:

`Run pip install -r pyproject.toml

Defaulting to user installation because normal site-packages is not writeable

ERROR: Invalid requirement: '[project]' (from line 1 of pyproject.toml)

Error: Process completed with exit code 1.`

I saw that other example .toml files had an initial '[build-system]' section, so I added that, but the error became:

`Run pip install -r pyproject.toml

Defaulting to user installation because normal site-packages is not writeable

ERROR: Invalid requirement: '[build-system]' (from line 1 of pyproject.toml)

Error: Process completed with exit code 1.`

Are you able to offer any advice? I was able to install the environment locally using PDM when I first created the .toml
