### 2024-04-02 6c976d1
reverted back to pip for ci
