### 2024-04-03 WSWUP/agweather-qaqc#24 @dostuffthatmatters
Hi @cwdunkerly,

apologies, I gave you the wrong command. `pip install .` installs the dependencies from the `pyproject.toml` file. The command  `pip install -r filename` only works for `requirements.txt` files. No need for `pdm-project/setup-pdm@v4`.

Also, in your current `pyproject.toml`, you should replace the line `requires-python = ">=3.9.*"` with `requires-python = ">=3.9.0,<4.0"`. The former doesn't work because `>=`/`<=` need exact versions, not stars.
