### 2024-04-03 #6368 @cwdunkerly
The last thing I have to do is make higher res versions of the figures. I'll have that done once I get out of my meetings this morning.

Do I need to reupload the paper anywhere when done or does JOSS pull it from the repo?
