### 2024-04-03 #6368 @cheginit
@cwdunkerly You can make all the necessary changes to your `paper.md` file, figures, and references in your repo, commit them, then use `editorialbot` to generate a new PDF (`@editorialbot generate pdf`).
