### 2024-04-04 7e93b71
made corrections to paper.md, added larger text to plots for readability
