+++ paper/paper.md
@@ -45,42 +45,41 @@ Agricultural weather station data are increasingly being used to support satelli
 
 Having the ability to easily read, visualize, review, flag, and potentially remove, fill, or adjust historical and real time weather data is necessary to advance research and applications focused on evapotranspiration and water use estimation at local to global scales. The development of the ``agweather-qaqc`` Python package is intended to enable rapid, thorough, and efficient weather data review and QAQC for daily weather data. The ``agweather-qaqc`` package is a command-line interface (CLI) based, open-source Python package for reading, visualizing, and performing QAQC of daily weather station observations and calculation of reference ET from a wide range of data networks. Many station networks use different storage formats and recorded variables. For example, a network might report humidity data as vapor pressure, dew point temperature, specific humidity, or maximum, minimum, and average relative humidity. As a result, data input variables, units, and conversions are configurable within ``agweather-qaqc`` to flexibly handle multiple common input variables and formats so that all input data, visualization, QAQC, and calculation of reference ET can be performed in a programmatic, consistent, and easily repeatable manner. Additionally, the CLI-based approach enables researchers and practitioners that are not overly proficient with Python to easily use and understand the software through the use of helpful reminders, prompts, and recommended parameters. While this package was written with a focus on agricultural stations, it can be used to QAQC any source of weather data, but care must be taken when making corrections or drawing conclusions about reference ET.
 
+# State of the Field
+``agweather-qaqc`` builds on the concepts and functionality of the Ref-ET Software [@allen2000], which performs QAQC of data and calculates ASCE reference ET. ``agweather-qaqc`` builds on these capabilities, is open source, streamlines the processing of data, and increases accessibility and batch processing by operating outside the Windows platform. Open source access also allows for the inclusion of additional methods and options. Other packages that have been developed, such as the R package [CrowdQC+](https://github.com/dafenner/CrowdQCplus) and the python library [titanlib](https://github.com/metno/titanlib), feature statistics-based QAQC of data, however they focus on spatial variation between multiple sources to detect outliers. ``agweather-qaqc`` differs from these libraries by focusing on the temporal variation within a single source.
+
+
+# Research Enabled by ``agweather-qaqc``
+This software was used to generate a benchmark agricultural weather station reference ET dataset for intercomparison and bias correction of a gridded weather dataset, gridMET [@abatzoglou2013] as part of OpenET [@melton2022], an online platform that provides satellite-based actual ET estimates at field scale across the western United States. This software has also been used to generate datasets of reference ET for weather stations in the Upper Colorado River Basin (UCRB) as part of a yearly Bureau of Reclamation publication of agricultural consumptive uses and losses [@pearson2019; @pearson2020; @pearson2021], intercomparison of OpenET models for the UCRB [@huntington2022], and was used in a skill analysis of NOAA’s Forecast Reference Evapotranspiration [@mcevoy2022].
+
 # Design and Features
 
 The ``agweather-qaqc`` package contains the WeatherQC python class that first reads and standardizes the data, and then allows the user to QAQC the data (\autoref{fig:fig1}).
 
-![Diagram showing the required inputs, flow of data, and outputs produced from the software.\label{fig:fig1}](figure1.pdf)
+![Diagram showing the required inputs, flow of data, and outputs produced from the software.\label{fig:fig1}](figure1.png)
 
 Inputs to ``agweather-qaqc`` are a tabular data file of daily data and a configuration file that details where variables are arranged within the data file. The configuration file also provides metadata about the station, such as latitude, longitude, elevation, and units of the variables. An optional input of a metadata file for multiple stations may also be provided, enabling a single configuration file to be used for any number of data sources that share a common format. 
 
[…]
