+++ paper/paper.md
@@ -43,10 +43,10 @@ Agricultural weather station data are increasingly being used to support satelli
 
 # Statement of Need
 
-Having the ability to easily read, visualize, review, flag, and potentially remove, fill, or adjust historical and real time weather data is necessary to advance research and applications focused on evapotranspiration and water use estimation at local to global scales. The development of the ``agweather-qaqc`` Python package is intended to enable rapid, thorough, and efficient weather data review and QAQC for daily weather data. The ``agweather-qaqc`` package is a command-line interface (CLI) based, open-source Python package for reading, visualizing, and performing QAQC of daily weather station observations and calculation of reference ET from a wide range of data networks. Many station networks use different storage formats and recorded variables. For example, a network might report humidity data as vapor pressure, dew point temperature, specific humidity, or maximum, minimum, and average relative humidity. As a result, data input variables, units, and conversions are configurable within ``agweather-qaqc`` to flexibly handle multiple common input variables and formats so that all input data, visualization, QAQC, and calculation of reference ET can be performed in a programmatic, consistent, and easily repeatable manner. Additionally, the CLI-based approach enables researchers and practitioners that are not overly proficient with Python to easily use and understand the software through the use of helpful reminders, prompts, and recommended parameters. While this package was written with a focus on agricultural stations, it can be used to QAQC any source of weather data, but care must be taken when making corrections or drawing conclusions about reference ET.
+Having the ability to easily read, visualize, review, flag, and potentially remove, fill, or adjust historical and real time weather data is necessary to advance research and applications focused on evapotranspiration and water use estimation at local to global scales. The development of the ``agweather-qaqc`` Python package is intended to enable rapid, thorough, and efficient weather data review and QAQC for daily weather data. The ``agweather-qaqc`` package is a command-line interface (CLI) based, open-source Python package for reading, visualizing, and performing QAQC of daily weather station observations and calculation of reference ET from a wide range of data networks. Many station networks use different storage formats and recorded variables. For example, a network might report humidity data as vapor pressure, dew point temperature, or maximum, minimum, and average relative humidity.  ``agweather-qaqc`` is able to flexibly handle multiple common input variables, units and formats so that all input data, visualization, QAQC, and calculation of reference ET can be performed in a programmatic, consistent, and easily repeatable manner. Additionally, the CLI-based approach enables researchers and practitioners that are not overly proficient with Python to easily use and understand the software through the use of helpful reminders, prompts, and recommended parameters. While this package was written with a focus on agricultural stations, it can be used to QAQC any source of weather data, but care must be taken when making corrections or drawing conclusions about reference ET.
 
 # State of the Field
-``agweather-qaqc`` builds on the concepts and functionality of the Ref-ET Software [@allen2000], which performs QAQC of data and calculates ASCE reference ET. ``agweather-qaqc`` builds on these capabilities, is open source, streamlines the processing of data, and increases accessibility and batch processing by operating outside the Windows platform. Open source access also allows for the inclusion of additional methods and options. Other packages that have been developed, such as the R package [CrowdQC+](https://github.com/dafenner/CrowdQCplus) and the python library [titanlib](https://github.com/metno/titanlib), feature statistics-based QAQC of data, however they focus on spatial variation between multiple sources to detect outliers. ``agweather-qaqc`` differs from these libraries by focusing on the temporal variation within a single source.
[…]
