### 2024-04-04 b72dcb5
Additional fixes to paper
