+++ paper/paper.md
@@ -68,13 +68,12 @@ Two optional functionalities can be enabled by the configuration file: The first
 
 One of the most useful features of the software is the interactive visualization of weather variable time series that facilitate rapid human data assessment and pattern recognition.The observation data are displayed in [Bokeh](https://github.com/bokeh/bokeh) [-@Bokeh] time series plots both before and after adjustment via the QAQC process, enabling users to readily visualize and assess data patterns and trends related to sensor drift, probable miscalibration, data outliers, sensor malfunction, etc. These plots are accompanied by interactive tools (ex. pan and zoom controls, display info on hover), and feature linked axes, allowing the user to readily visualize and assess how variables vary and covary over time and at different time scales (e.g. daily, monthly, or annual). These graphs are saved as stand-alone HTML files to allow for easy sharing of results.
 
-The interactive plots also facilitate viewing data patterns at different timescales, such as showing how minimum temperature and dew point temperature vary at daily and monthly timesteps, as well as how the difference in these variables, known as dew point depression, varies throughout the year. Dew point depression gives insight to how well-watered the area around the station may be: transpiration from plants and evaporation from soil increases humidity, which results in increased dew point and reduced dew point depression. Higher than expected or abrupt changes in dew point depression values might indicate non-ideal reference ET station siting, land cover changes, or erroneous data.
+The interactive plots also facilitate viewing patterns in the data, such as showing how minimum temperature and dew point temperature vary at daily and monthly timesteps, as well as how the difference in these variables, known as dew point depression, varies throughout the year. Dew point depression gives insight to how well-watered the area around the station may be: transpiration from plants and evaporation from soil increases humidity, which results in increased dew point and reduced dew point depression. Higher than expected or abrupt changes in dew point depression values might indicate non-ideal reference ET station siting, land cover changes, or erroneous data.
 
 In addition to already described process, ``agweather-qaqc`` also performs the following steps:
 
 * Calculation of theoretical clear-sky solar radiation using date, location and elevation information along with humidity data based on ASCE standardizations [@asce-ewri2005].
 * Calculation of expected solar radiation using the empirical Thornton-Running approach [@thornton1999] with Monte-Carlo optimized empirical parameters based on observed solar radiation data.
-* Calculation of the daily dew point depression (i.e., daily minimum temperature minus daily average dew point temperature) used to assess whether the data collection environment included a well-watered surface having expected feedbacks on near-surface humidity and air temperature.
 * Calculation of grass and alfalfa reference evapotranspiration according to the American Society of Civil Engineers (ASCE) Standardized Reference Evapotranspiration equation [@asce-ewri2005] via the [RefET](https://github.com/WSWUP/RefET) [@RefET] library.
 * Evaluating station aridity through the visualization of both relative humidity and dew point depression plots, with the option to adjust relative humidity if required.
 * Creating archival charts and log files that record and flag how each variable was changed during the QAQC process.
@@ -89,7 +88,7 @@ Inaccuracies in measured incoming shortwave solar radiation (R~s~) are common du
 
 It is expected that a cloud-free day should occur with some regularity for areas with significant agriculture, especially in semi-arid and arid areas. In this example time series (\autoref{fig:fig2}), observed R~s~ from before July 2013 frequently approaches R~so~, and after this time period, R~s~ rarely approaches R~so~, especially during summer and winter months, indicating that the pyranometer is likely reporting inaccurate values and that observed data should be adjusted.
 
[…]
