### 2024-04-04 20e8ebb
formatted a paragraph in paper.md
