+++ paper/paper.md
@@ -88,7 +88,7 @@ Inaccuracies in measured incoming shortwave solar radiation (R~s~) are common du
 
 It is expected that a cloud-free day should occur with some regularity for areas with significant agriculture, especially in semi-arid and arid areas. In this example time series (\autoref{fig:fig2}), observed R~s~ from before July 2013 frequently approaches R~so~, and after this time period, R~s~ rarely approaches R~so~, especially during summer and winter months, indicating that the pyranometer is likely reporting inaccurate values and that observed data should be adjusted.
 
-![(Top) Observed shortwave radiation compared against expected clear-sky solar radiation for a weather station in Reno, Nevada. The downward drift in the sensor readings over the four year period is made obvious by comparing against the theoretical clear-sky values. (Middle) Post-QC shortwave radiation compared against clear-sky solar radiation, the downward drift has been adjusted back to the expected behavior of R~s~. (Bottom) The percent difference between pre- and post-QC R~s~ values as a result of applying the $\frac{R~so~}{R~s~}$ correction factor ratio to each sixty-day period..\label{fig:fig2}](figure2.png)
+![(Top) Observed shortwave radiation compared against expected clear-sky solar radiation for a weather station in Reno, Nevada. The downward drift in the sensor readings over the four year period is made obvious by comparing against the theoretical clear-sky values. (Middle) Post-QC shortwave radiation compared against clear-sky solar radiation, the downward drift has been adjusted back to the expected behavior of R~s~. (Bottom) The percent difference between pre- and post-QC R~s~ values as a result of applying the $\frac{R~so~}{R~s~}$ correction factor ratio to each sixty-day period.\label{fig:fig2}](figure2.png)
 
 To perform the adjustment, measured R~s~ data are parsed into 60-day periods. Each period is first checked for [bad or suspect values](https://wswup.github.io/agweather-qaqc/agweatherqaqc.html#agweatherqaqc.qaqc_functions.rs_period_ratio_corr), and then the ratio of R~so~ to R~s~ is calculated from the average of the six largest daily R~s~ observations and corresponding R~so~ values (i.e. top 10 percentile for the 60-day period). This ratio is multiplied by all R~s~ measurements during the respective 60-day period to correct measured R~s~ data. If the R~so~ to R~s~ ratio is between 0.97 and 1.03, no adjustment is applied. The length of period and number of largest R~s~/R~so~ ratios (e.g. percentile threshold) to compute average adjustment ratios for the respective period are configurable through interactive prompts.
 
