### 2024-04-04 WSWUP/agweather-qaqc#19 @cwdunkerly
Responding to individual points:

1: Done, if you'd like them to be even higher resolution please let me know.
3: I went over the duplicate text you mentioned and identified some more myself, I have trimmed it down appropriately.
4: I changed this plot to have full titles and then referenced the different components by their relative positions.
6: Thank you for the reminder on grammar, I did find a few and fixed them.
7: I added the section you requested. I personally feel this info might be better documented in the repo as opposed to the paper proper, but if you feel differently I understand.
