### 2024-04-10 WSWUP/agweather-qaqc#28 @dlebauer
[Feature Request] include reproducible results from JOSS paper fig 2
**Please describe the feature you'd like to see**

JOSS Requires 'reproducibility' defined as

> If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

Please add the example data and script or README to reproduce the analysis in the "paper" folder or subdirectory thereof.
