### 2024-04-10 #6368 @dlebauer
@cwdunkerly this is a nice paper and I appreciate all of the improvements that you have made in response to feedback. Well done, almost there! The only thing remaining is to make sure that data, code, and instructions to reproduce the figure 2 are provided https://github.com/WSWUP/agweather-qaqc/issues/28
