### 2024-04-11 WSWUP/agweather-qaqc#19 @dostuffthatmatters
@cwdunkerly Thank you for implementing the changes!

Regarding 7: You can also move it to the documentation if you like that better. Having it in the paper is more permanent though. I see that in more and more publication but it is indeed not a standardised thing yet. (ttps://www.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement)
