### 2024-04-11 WSWUP/agweather-qaqc#19 @dostuffthatmatters
@cwdunkerly

I opened one more PR (#29) to add more metadata (dois, etc.) to the bibliography.
