### 2024-04-11 #6368 @dostuffthatmatters
I have the same impression as @dlebauer. The tool is well done and the repository has improved a lot in the last weeks. I only see two things from my end:

- [ ] I opened a PR because there was a lot of metadata that could be added to the bibliography: https://github.com/WSWUP/agweather-qaqc/pull/29
- [ ] The installation instructions should be updated to the ones from the README

After that, I recommend it to be accepted. Good work, @cwdunkerly!
