### 2024-04-11 #6368 @cwdunkerly
Thank you both very much for your supportive comments. I do have one question:

@dlebauer I had used the CIMIS weather station located in Davis, CA for my example files as it's a very well documented station, but the data used for figure 2 is from a station in Reno, NV. Would it be satisfactory if I was to include this Reno station as "example_data_3" and include text in the figure description detailing which file it came from?

@cheginit I am out of the office until Tuesday morning, I will have these final comments resolved and pushed that day, and I'll send a message when it's done.
