### 2024-04-12 #6368 @dlebauer
> Would it be satisfactory if I was to include this Reno station as "example_data_3" and include text in the figure description detailing which file it came from?

Outside of a JOSS submission I think that would be sufficient. But given my understanding of JOSS’ reproducibility requirement, I think the data is insufficient, there should be code and instructions for reproducing the figure. This could be in a simple md file alongside the dataset.
