### 2024-04-24 WSWUP/agweather-qaqc#19 @dostuffthatmatters
> @dostuffthatmatters Not going to reopen this but one of the things I am finishing today is changing the author credit to use the terms in the link you shared. Thank you for providing that link.

Oh, sorry for closing it too quickly ^^
