### 2024-04-24 WSWUP/agweather-qaqc#19 @cwdunkerly
No need to apologize, the fault is mine for leaving this open.
