### 2024-04-25 0740354
updated installation instructions in docs to match what was in readme
