### 2024-04-25 b5ebc29
added data and instructions to recreate figure 2 of the joss paper
