+++ tests/test_files/joss_figure_2/joss_figure_2_config.ini
@@ -1,0 +1,157 @@
+############################################################################################################################
+############################################################################################################################
+[METADATA]
+# This section is used to specify information about the weather station and the data file being processed.
+
+
+# SPECIFY THE PATH TO THE DATA FILE (MUST BE .csv, .xls, OR .xlsx)
+DATA_FILE_PATH = tests/test_files/joss_figure_2/joss_figure_2_data.csv
+
+
+# STATION LATITUDE AND LONGITUDE (IN DECIMAL DEGREES)
+LATITUDE = 39.503056
+LONGITUDE = -119.737778
+
+
+# STATION ELEVATION (IN METERS)
+ELEVATION = 1338.072
+
+
+# STATION ANEMOMETER HEIGHT ABOVE GROUND (IN METERS)
+ANEMOMETER_HEIGHT = 3
+
+
+# VALUES TO SIGNIFY MISSING OBSERVATIONS IN INPUT AND OUTPUT FILES
+#	EX. -999, 12345, NO_DATA
+MISSING_INPUT_VALUE = -9999
+MISSING_OUTPUT_VALUE = nan
+
+
+# LINES OF HEADER AND FOOTER IN INPUT FILE TO SKIP READING OVER (STARTING AT INDEX 1)
+#	EX: IF DATA STARTS ON LINE 3 OF FILE, SET LINES_OF_HEADER = 2
+#	IF NO HEADER OR FOOTER EXISTS SET TO 0
+LINES_OF_HEADER = 1
+LINES_OF_FOOTER = 0
+
+############################################################################################################################
+############################################################################################################################
+[OPTIONS]
+# This section decides which options the script will use when it is run. Each option is explained in more detail below:
+
+
+# AUTOMATED OPTION - THE SCRIPT WILL DO THE RECCOMMENDED COURSE OF ACTION ON EACH VARIABLE YOU SELECT
+#	THIS MAY CAUSE MORE WORK FOR THE USER IF THE DATA REQUIRES A LOT OF CORRECTION
+#	0 - OFF
+#	1 - ON
+AUTOMATIC_OPTION = 0
+
+
+# FILLING OPTION - THE SCRIPT WILL FILL IN ALL MISSING DATA TO END WITH A COMPLETE RECORD OF REFERENCE EVAPOTRANSPIRATION
+# 	NOT RECOMMENDED IF GOING FOR ACCURACY, HOW VARIABLES ARE FILLED IS DETAILED IN THE DOCUMENTATION
+#	0 - DO NOT FILL MISSING DATA
+#	1 - FILL MISSING DATA
+FILL_OPTION = 0
+
+
+# OUTPUT FILE FORMAT FOR CORRECTED DATA - MUST BE EITHER 'CSV' OR 'XLSX'
+OUTPUT_DATA_FORMAT = XLSX
+
+
+############################################################################################################################
+############################################################################################################################
+[DATA]
+# This section specifies which columns variables are in, as well as what units they are provided in.
[…]
