### 2024-04-25 WSWUP/agweather-qaqc#28 @cwdunkerly
I have included the files necessary to re-create the figure and instructions in a markdown file of the steps to go through. They are located here: https://github.com/WSWUP/agweather-qaqc/tree/master/tests/test_files/joss_figure_2

Please let me know if this is insufficient, I am not going to close this until I have your sign off.
