### 2024-04-25 #6368 @cwdunkerly
I have completed all outstanding tasks, sorry for the delay. If I have missed something please let me know.

@dostuffthatmatters I have updated the docs with installation instructions, and I changed the language for the contributor roles in the paper, if you wished to look over that as well.

@dlebauer I have added instructions for re-creating figure 2, as requested under [28](https://github.com/WSWUP/agweather-qaqc/issues/28). If you think more is needed please let me know.

@cheginit the final thing I wish to do is review the paper one final time later today. I'm catching a flight this afternoon and I plan to do that then.
