+++ paper/paper.md
@@ -62,7 +62,7 @@ Inputs to ``agweather-qaqc`` are a tabular data file of daily data and a configu
 
 The ``agweather-qaqc`` package contains [example files](https://github.com/WSWUP/agweather-qaqc/tree/master/tests/test_files) of each input.
 
-Once given data, the WeatherQC class performs unit conversions and removal of unreasonable values according to limits and ranges specified in the [documentation](https://wswup.github.io/agweather-qaqc/), (e.g., negative wind speed readings). This is followed by calculating variables, including ASCE reference ET, mean monthly averages, and other diagnostic metrics. Command-line interface (CLI) prompts facilitate corrections on desired variables, and provides recommendations as to best practices. All downstream dependent variables are recalculated once each correction is performed to assemble the quality-controlled version of the data. Once the user is finished making adjustments, output files are generated that include records of all changes through both a human-readable log file and interactive time series plots of pre- and post-processed observations. The generated output files form a complete record of original and processed observations with correction flags and amounts and station metadata so that a documented data archive is available for later use.
+Once given data, the WeatherQC class performs unit conversions and removal of unreasonable values according to limits and ranges specified in the [documentation](https://wswup.github.io/agweather-qaqc/qc_information.html), (e.g., negative wind speed readings). This is followed by calculating variables, including ASCE reference ET, mean monthly averages, and other diagnostic metrics. Command-line interface (CLI) prompts facilitate corrections on desired variables, and provides recommendations as to best practices. All downstream dependent variables are recalculated once each correction is performed to assemble the quality-controlled version of the data. Once the user is finished making adjustments, output files are generated that include records of all changes through both a human-readable log file and interactive time series plots of pre- and post-processed observations. The generated output files form a complete record of original and processed observations with correction flags and amounts and station metadata so that a documented data archive is available for later use.
 
 Two optional functionalities can be enabled by the configuration file: The first automates some of the interactive prompts to use the suggested QC procedures and parameters at each step. The second provides options for gap-filling using station climatologies, empirical approaches (e.g. Thornton-Running solar), or random sampling from observations, to provide a complete record of daily reference evapotranspiration for cumulative monthly and annual totals.
 
