+++ paper/paper.bib
@@ -32,7 +32,7 @@
   number = {2},
   journal = {Journal of Irrigation and Drainage Engineering},
   publisher = {American Society of Civil Engineers (ASCE)},
-  author = {Allen,  Richard G.},
+  author = {Allen, Richard G.},
   year = {1996},
   month = mar,
   pages = {97–106}
@@ -109,7 +109,7 @@
 @article{huntington2022,
   doi = {10.13140/RG.2.2.21605.88808},
   url = {https://rgdoi.net/10.13140/RG.2.2.21605.88808},
-  author = {Huntington,  Justin L and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles and Melton,  Forrest and Allen,  Richard},
+  author = {Huntington,  Justin L. and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles and Melton,  Forrest and Allen,  Richard},
   language = {en},
   title = {Upper Colorado River Basin OpenET Intercomparison Summary},
   publisher = {Desert Research Institute},
