+++ paper/paper.bib
@@ -32,7 +32,7 @@
   number = {2},
   journal = {Journal of Irrigation and Drainage Engineering},
   publisher = {American Society of Civil Engineers (ASCE)},
-  author = {Allen, Richard G.},
+  author = {Allen, Richard},
   year = {1996},
   month = mar,
   pages = {97–106}
@@ -47,24 +47,9 @@
   number = {4},
   journal = {Journal of Agricultural Meteorology},
   publisher = {Society of Agricultural Meteorology of Japan},
-  author = {ALLEN,  Richard},
+  author = {Allen,  Richard},
   year = {2008},
   pages = {191–204}
-}
-
-@article{allen2011,
-  title = {Evapotranspiration information reporting: I. Factors governing measurement accuracy},
-  volume = {98},
-  ISSN = {0378-3774},
-  url = {https://www.sciencedirect.com/science/article/abs/pii/S0378377411000023},
-  DOI = {10.1016/j.agwat.2010.12.015},
-  number = {6},
-  journal = {Agricultural Water Management},
-  publisher = {Elsevier BV},
-  author = {Allen,  Richard G. and Pereira,  Luis S. and Howell,  Terry A. and Jensen,  Marvin E.},
-  year = {2011},
-  month = apr,
-  pages = {899–920}
 }
 
 @article{allen2021,
@@ -75,7 +60,7 @@
   DOI = {10.1016/j.agwat.2020.106531},
   journal = {Agricultural Water Management},
   publisher = {Elsevier BV},
-  author = {Allen,  Richard G. and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin and Kilic,  Ayse and Morton,  Charles},
+  author = {Allen,  Richard and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin and Kilic,  Ayse and Morton,  Charles},
   year = {2021},
   month = feb,
   pages = {106531}
@@ -222,7 +207,7 @@
   url = {https://ascelibrary.org/doi/book/10.1061/9780784408056},
   DOI = {10.1061/9780784408056},
   publisher = {American Society of Civil Engineers},
-  editor = {Allen,  Richard G. and Walter,  Ivan A. and Elliott,  Ronald L. and Howell,  Terry A. and Itenfisu,  Daniel and Jensen,  Marvin E. and Snyder,  Richard L.},
+  editor = {Allen,  Richard and Walter,  Ivan and Elliott,  Ronald and Howell,  Terry and Itenfisu,  Daniel and Jensen,  Marvin and Snyder,  Richard},
   year = {2005},
   month = aug 
 }
