### 2024-04-26 49072db
standardized names across bibliography
