+++ paper/paper.bib
@@ -15,7 +15,7 @@
 
 @article{allen1998,
   title = {Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
-  author = {Allen, R and Pereira, L and Raes, D and Smith, M and others},
+  author = {Allen, R. G. and Pereira, L and Raes, D and Smith, M and others},
   journal = {Fao, Rome},
   volume = {300},
   number = {9},
@@ -32,7 +32,7 @@
   number = {2},
   journal = {Journal of Irrigation and Drainage Engineering},
   publisher = {American Society of Civil Engineers (ASCE)},
-  author = {Allen, Richard},
+  author = {Allen, Richard G.},
   year = {1996},
   month = mar,
   pages = {97–106}
@@ -47,7 +47,7 @@
   number = {4},
   journal = {Journal of Agricultural Meteorology},
   publisher = {Society of Agricultural Meteorology of Japan},
-  author = {Allen,  Richard},
+  author = {Allen,  Richard G.},
   year = {2008},
   pages = {191–204}
 }
@@ -60,7 +60,7 @@
   DOI = {10.1016/j.agwat.2020.106531},
   journal = {Agricultural Water Management},
   publisher = {Elsevier BV},
-  author = {Allen,  Richard and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin and Kilic,  Ayse and Morton,  Charles},
+  author = {Allen,  Richard G. and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin and Kilic,  Ayse and Morton,  Charles},
   year = {2021},
   month = feb,
   pages = {106531}
@@ -207,7 +207,7 @@
   url = {https://ascelibrary.org/doi/book/10.1061/9780784408056},
   DOI = {10.1061/9780784408056},
   publisher = {American Society of Civil Engineers},
-  editor = {Allen,  Richard and Walter,  Ivan and Elliott,  Ronald and Howell,  Terry and Itenfisu,  Daniel and Jensen,  Marvin and Snyder,  Richard},
+  editor = {Allen,  Richard G. and Walter,  Ivan and Elliott,  Ronald and Howell,  Terry and Itenfisu,  Daniel and Jensen,  Marvin and Snyder,  Richard},
   year = {2005},
   month = aug 
 }
@@ -231,7 +231,7 @@ url = {https://bokeh.pydata.org/en/latest/},
 
 @misc{allen2000,
   title = {REF-ET: Reference evapotranspiration calculation software for FAO and ASCE standardized equations},
-  author = {Allen, R},
+  author = {Allen, Richard G.},
   journal = {University of Idaho},
   year = {2000},
   url = {https://www.uidaho.edu/cals/kimberly-research-and-extension-center/research/water-resources/ref-et-software},
