### 2024-04-29 e209b00
grammatical edit to paper.md
