+++ paper/paper.md
@@ -90,7 +90,7 @@ It is expected that a cloud-free day should occur with some regularity for areas
 
 ![(Top) Observed shortwave radiation compared against expected clear-sky solar radiation for a weather station in Reno, Nevada. The downward drift in the sensor readings over the four year period is made obvious by comparing against the theoretical clear-sky values. (Middle) Post-QC shortwave radiation compared against clear-sky solar radiation, the downward drift has been adjusted back to the expected behavior of R~s~. (Bottom) The percent difference between pre- and post-QC R~s~ values as a result of applying the $\frac{R~so~}{R~s~}$ correction factor ratio to each sixty-day period.\label{fig:fig2}](figure2.png)
 
-To perform the adjustment, measured R~s~ data are parsed into 60-day periods. Each period is first checked for [bad or suspect values](https://wswup.github.io/agweather-qaqc/agweatherqaqc.html#agweatherqaqc.qaqc_functions.rs_period_ratio_corr), and then the ratio of R~so~ to R~s~ is calculated from the average of the six largest daily R~s~ observations and corresponding R~so~ values (i.e. top 10 percentile for the 60-day period). This ratio is multiplied by all R~s~ measurements during the respective 60-day period to correct measured R~s~ data. If the R~so~ to R~s~ ratio is between 0.97 and 1.03, no adjustment is applied. The length of period and number of largest R~s~/R~so~ ratios (e.g. percentile threshold) to compute average adjustment ratios for the respective period are configurable through interactive prompts.
+To perform the adjustment, measured R~s~ data are parsed into 60-day periods. Each period is first checked for [bad or suspect values](https://wswup.github.io/agweather-qaqc/agweatherqaqc.html#agweatherqaqc.qaqc_functions.rs_period_ratio_corr), and then the ratio of R~so~ to R~s~ is calculated from the average of the six largest daily R~s~ observations and corresponding R~so~ values (i.e. top 10th percentile for the 60-day period). This ratio is multiplied by all R~s~ measurements during the respective 60-day period to correct measured R~s~ data. If the R~so~ to R~s~ ratio is between 0.97 and 1.03, no adjustment is applied. The length of period and number of largest R~s~/R~so~ ratios (e.g. percentile threshold) to compute average adjustment ratios for the respective period are configurable through interactive prompts.
 
 After adjustment of R~s~ is performed, a time series plot of percent difference between pre- and post-corrected R~s~ is produced for visualization and archiving adjustments made. If faulty R~s~ observations had not been adjusted, reference ET would be underestimated in this case by approximately 5%. This example highlights how ``agweather-qaqc`` can be used to visualize and correct weather variables to improve evaporative demand estimates.
 
