### 2024-04-29 #6368 @cwdunkerly
Other than @dlebauer confirming whether [28](https://github.com/WSWUP/agweather-qaqc/issues/28) is sufficiently resolved, I have two minor questions regarding the paper:

1. For line 153 in the article proof, is "top 10 percentile" or "top 10th percentile" more grammatically correct? The paper is currently "10th". I changed it after editorialbot generated the paper this most recent time.

2. Richard Allen is the only name in the in-paper citations that gets spelled out or has initials included. Rick is a fantastic guy and he's been a great teacher/mentor, but it does feel weird to me that the only citations like this happen to be for a name on the paper. Is there an easy way to change this beyond editing the bibliography directly?

Thanks.
