### 2024-05-06 e02567f
added ORCiD's and and urls for some references without DOIs
