+++ paper/paper.bib
@@ -14,7 +14,8 @@
 }
 
 @article{allen1998,
-  title = {Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
+  title = {Crop Evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
+  url = {https://www.researchgate.net/publication/290988983},
   author = {Allen, R. G. and Pereira, L and Raes, D and Smith, M and others},
   journal = {Fao, Rome},
   volume = {300},
@@ -180,6 +181,7 @@
 
 @article{steduto2012,
   title = {Coping with water scarcity: an action framework for agriculture and food security},
+  url = {https://openknowledge.fao.org/handle/20.500.14283/i3015e},
   author = {Steduto, Pasquale and Faur{\`e}s, Jean-Marc and Hoogeveen, Jippe and Winpenny, Jim and Burke, Jacob},
   journal = {Food and Agriculture Organization of the United Nations},
   year = {2012}
@@ -223,10 +225,10 @@
 }
 
 @Manual{Bokeh,
-title = {Bokeh: Python library for interactive visualization},
-author = {Bokeh Development Team},
-year = {2018},
-url = {https://bokeh.pydata.org/en/latest/},
+  title = {Bokeh: Python library for interactive visualization},
+  author = {Bokeh Development Team},
+  year = {2018},
+  url = {https://bokeh.pydata.org/en/latest/},
 }
 
 @misc{allen2000,
