### 2024-05-06 4e75c3e
standardized author names across references and in paper
