+++ paper/paper.bib
@@ -16,7 +16,7 @@
 @article{allen1998,
   title = {Crop Evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
   url = {https://www.researchgate.net/publication/290988983},
-  author = {Allen, R. G. and Pereira, L and Raes, D and Smith, M and others},
+  author = {Allen, Richard G. and Pereira, L and Raes, D and Smith, M and others},
   journal = {Fao, Rome},
   volume = {300},
   number = {9},
@@ -61,7 +61,7 @@
   DOI = {10.1016/j.agwat.2020.106531},
   journal = {Agricultural Water Management},
   publisher = {Elsevier BV},
-  author = {Allen,  Richard G. and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin and Kilic,  Ayse and Morton,  Charles},
+  author = {Allen,  Richard G. and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin L. and Kilic,  Ayse and Morton,  Charles},
   year = {2021},
   month = feb,
   pages = {106531}
@@ -75,7 +75,7 @@
   DOI = {10.1016/j.agwat.2020.106376},
   journal = {Agricultural Water Management},
   publisher = {Elsevier BV},
-  author = {Blankenau,  Philip A. and Kilic,  Ayse and Allen,  Richard},
+  author = {Blankenau,  Philip A. and Kilic,  Ayse and Allen,  Richard G.},
   year = {2020},
   month = dec,
   pages = {106376}
@@ -95,7 +95,7 @@
 @article{huntington2022,
   doi = {10.13140/RG.2.2.21605.88808},
   url = {https://rgdoi.net/10.13140/RG.2.2.21605.88808},
-  author = {Huntington,  Justin L. and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles and Melton,  Forrest and Allen,  Richard},
+  author = {Huntington,  Justin L. and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles and Melton,  Forrest and Allen,  Richard G.},
   language = {en},
   title = {Upper Colorado River Basin OpenET Intercomparison Summary},
   publisher = {Desert Research Institute},
@@ -140,7 +140,7 @@
   number = {6},
   journal = {JAWRA Journal of the American Water Resources Association},
   publisher = {Wiley},
-  author = {Melton,  Forrest S. and Huntington,  Justin and Grimm,  Robyn and Herring,  Jamie and Hall,  Maurice and Rollison,  Dana and Erickson,  Tyler and Allen,  Richard and Anderson,  Martha and Fisher,  Joshua B. and Kilic,  Ayse and Senay,  Gabriel B. and Volk,  John and Hain,  Christopher and Johnson,  Lee and Ruhoff,  Anderson and Blankenau,  Philip and Bromley,  Matt and Carrara,  Will and Daudert,  Britta and Doherty,  Conor and Dunkerly,  Christian and Friedrichs,  MacKenzie and Guzman,  Alberto and Halverson,  Gregory and Hansen,  Jody and Harding,  Jordan and Kang,  Yanghui and Ketchum,  David and Minor,  Blake and Morton,  Charles and Ortega‐Salazar,  Samuel and Ott,  Thomas and Ozdogan,  Mutlu and ReVelle,  Peter M. and Schull,  Mitch and Wang,  Carlos and Yang,  Yun and Anderson,  Ray G.},
[…]
