### 2024-05-06 323b6bd
fixed co-author roles in paper.md
