### 2024-05-06 #6368 @cwdunkerly
@cheginit  Sorry for the delay, I had several field trips last week. I have done everything you asked, to recap:
- Author names have been standardized across references.
- I have added ORCiDs for the co-authors that have one.
- I have confirmed that the references with missing DOIs do not have them.
	- I have added URL's to the bibliography for two sources without DOIs.
- I have made a tagged release of the software ([v1.0.4](https://github.com/WSWUP/agweather-qaqc/releases/tag/v1.0.4))
- I have archived the software (in the form of a .zip of the above tagged release) to [Zenodo](https://zenodo.org/records/11122799).
	- The software license matches the github repo
	- The metadata matches the paper draft
	- The DOI is 10.5281/zenodo.11122799
