### 2024-05-07 #6368 @kthyng
Paper comments:
* When you say "well-watered" do you mean an area is watered from a well or it has been watered sufficiently? I suppose someone in the field would know this but I couldn't tell from reading the paper.
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
