### 2024-05-10 #6368 @cwdunkerly
@kthyng Thank you for your comments, I will get these resolved.

Well-watered means the latter, I will change the paper text and review the bibliography today and notify you when it is done.

Since I am changing the paper, do I also need to update the files on zenodo?
