### 2024-05-10 #6368 @kthyng
> Since I am changing the paper, do I also need to update the files on zenodo?
No, the generated paper acts as the archive of the paper itself and the code hasn't changed.
