+++ paper/paper.bib
@@ -112,7 +112,7 @@
   number = {1},
   journal = {Scientific Reports},
   publisher = {Springer Science and Business Media LLC},
-  author = {Kummu,  M. and Guillaume,  J. H. A. and de Moel,  H. and Eisner,  S. and Fl\"{o}rke,  M. and Porkka,  M. and Siebert,  S. and Veldkamp,  T. I. E. and Ward,  P. J.},
+  author = {Kummu,  M. and Guillaume,  J. H. A. and {de Moel},  H. and Eisner,  S. and Fl\"{o}rke,  M. and Porkka,  M. and Siebert,  S. and Veldkamp,  T. I. E. and Ward,  P. J.},
   year = {2016},
   month = dec 
 }
@@ -226,7 +226,7 @@
 
 @Manual{Bokeh,
   title = {Bokeh: Python library for interactive visualization},
-  author = {Bokeh Development Team},
+  author = {{Bokeh Development Team}},
   year = {2018},
   url = {https://bokeh.pydata.org/en/latest/},
 }
