+++ paper/paper.bib
@@ -16,7 +16,7 @@
 @article{allen1998,
   title = {Crop Evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
   url = {https://www.researchgate.net/publication/290988983},
-  author = {Allen, Richard G. and Pereira, L and Raes, D and Smith, M and others},
+  author = {Allen, Richard G. and Pereira, L and Raes, D and Smith, M},
   journal = {Fao, Rome},
   volume = {300},
   number = {9},
@@ -61,7 +61,7 @@
   DOI = {10.1016/j.agwat.2020.106531},
   journal = {Agricultural Water Management},
   publisher = {Elsevier BV},
-  author = {Allen,  Richard G. and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin L. and Kilic,  Ayse and Morton,  Charles},
+  author = {Allen,  Richard G. and Dhungel,  Ramesh and Dhungana,  Bibha and Huntington,  Justin L. and Kilic,  Ayse and Morton,  Charles G.},
   year = {2021},
   month = feb,
   pages = {106531}
@@ -95,7 +95,7 @@
 @article{huntington2022,
   doi = {10.13140/RG.2.2.21605.88808},
   url = {https://rgdoi.net/10.13140/RG.2.2.21605.88808},
-  author = {Huntington,  Justin L. and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles and Melton,  Forrest and Allen,  Richard G.},
+  author = {Huntington,  Justin L. and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles G. and Melton,  Forrest and Allen,  Richard G.},
   language = {en},
   title = {Upper Colorado River Basin OpenET Intercomparison Summary},
   publisher = {Desert Research Institute},
@@ -216,7 +216,7 @@
 
 
 @misc{RefET,
-  author = {Morton, C},
+  author = {Morton, Charles G.},
   title = {ASCE Standardized Reference Evapotranspiration (ET)},
   year = {2022},
   publisher = {GitHub},
