+++ paper/paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 
 # Introduction
 
-Agricultural weather stations are necessary for collecting representative observations needed for computing reference evapotranspiration (ET), a measure of atmospheric evaporative demand from a sufficiently-watered environment, and a key variable for determining crop water use, water requirements, and irrigation scheduling [@allen1998]. Common weather station measurements include near-surface (2 to 3 m height) measurements for incoming shortwave solar radiation, air temperature, humidity, wind speed, and precipitation. Accurate, continuous, and consistent measurement of these variables from years to decades is often lacking due to sensor malfunction, drift, age, poor calibration, debris, limited station maintenance, communication errors, lack of a sufficiently-watered environment, and remote access. As a result, poor quality agricultural weather station data are common, and, unless flagged for removal or corrected, will impact the accuracy of reference ET calculations [@allen1996]. Ensuring that agricultural station data are high quality is especially important for studies where station data are considered to be ‘truth’ when comparing to related model predictions and forecasts [@blankenau2020; @mcevoy2022]. Similarly, reference ET station data need to represent sufficiently-watered environments if they are to be used to validate or to bias-correct gridded reference ET datasets [@allen2021; @huntington2018]. 
+Agricultural weather stations are necessary for collecting representative observations needed for computing reference evapotranspiration (ET), a measure of atmospheric evaporative demand from a sufficiently watered environment, and a key variable for determining crop water use, water requirements, and irrigation scheduling [@allen1998]. Common weather station measurements include near-surface (2 to 3 m height) measurements for incoming shortwave solar radiation, air temperature, humidity, wind speed, and precipitation. Accurate, continuous, and consistent measurement of these variables from years to decades is often lacking due to sensor malfunction, drift, age, poor calibration, debris, limited station maintenance, communication errors, lack of a sufficiently watered environment, and remote access. As a result, poor quality agricultural weather station data are common, and, unless flagged for removal or corrected, will impact the accuracy of reference ET calculations [@allen1996]. Ensuring that agricultural station data are high quality is especially important for studies where station data are considered to be ‘truth’ when comparing to related model predictions and forecasts [@blankenau2020; @mcevoy2022]. Similarly, reference ET station data need to represent sufficiently watered environments if they are to be used to validate or to bias-correct gridded reference ET datasets [@allen2021; @huntington2018]. 
 
 Agricultural weather station data are increasingly being used to support satellite-based remote sensing of agricultural water use. The weather data are often collected from dozens of networks and hundreds of stations and requires large-scale quality assurance and quality control (QAQC) systems to ensure accuracy and representativeness [@melton2021; @huntington2022]. A common challenge has been the development of QAQC workflows and visualization tools to support implementation of recent advances in recognizing and correcting data error. Reliable, high-quality records of weather observations will become even more important as scarcity in freshwater resources continues to become a more prevalent issue in the world [@steduto2012; @kummu2016].
 
@@ -70,7 +70,7 @@ Two optional functionalities can be enabled by the configuration file: The first
 
[…]
