### 2024-05-13 85a59ba
Fixed capitalization of titles in bibliography
