+++ paper/paper.bib
@@ -14,10 +14,10 @@
 }
 
 @article{allen1998,
-  title = {Crop Evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
+  title = {{Crop Evapotranspiration-guidelines for computing crop water requirements-FAO Irrigation and Drainage Paper No. 56}},
   url = {https://www.researchgate.net/publication/290988983},
   author = {Allen, Richard G. and Pereira, L and Raes, D and Smith, M},
-  journal = {Fao, Rome},
+  journal = {Food and Agriculture Organization of the United Nations},
   volume = {300},
   number = {9},
   pages = {D05109},
@@ -25,7 +25,7 @@
 }
 
 @article{allen1996,
-  title = {Assessing Integrity of Weather Data for Reference Evapotranspiration Estimation},
+  title = {{Assessing Integrity of Weather Data for Reference Evapotranspiration Estimation}},
   volume = {122},
   ISSN = {1943-4774},
   url = {https://ascelibrary.org/doi/10.1061/%28ASCE%290733-9437%281996%29122%3A2%2897%29},
@@ -40,7 +40,7 @@
 }
 
 @article{allen2008,
-  title = {Quality Assessment of Weather Data and Micrometeological Flux-Impacts on Evapotranspiration Calculation},
+  title = {{Quality Assessment of Weather Data and Micrometeological Flux-Impacts on Evapotranspiration Calculation}},
   volume = {64},
   ISSN = {1881-0136},
   url = {https://www.jstage.jst.go.jp/article/agrmet/64/4/64_64.4.5/_article},
@@ -68,7 +68,7 @@
 }
 
 @article{blankenau2020,
-  title = {An evaluation of gridded weather data sets for the purpose of estimating reference evapotranspiration in the United States},
+  title = {An evaluation of gridded weather data sets for the purpose of estimating reference evapotranspiration in the {U}nited {S}tates},
   volume = {242},
   ISSN = {0378-3774},
   url = {https://www.sciencedirect.com/science/article/abs/pii/S0378377420304753},
@@ -86,7 +86,7 @@
   url = {http://rgdoi.net/10.13140/RG.2.2.35811.60967},
   author = {Huntington,  Justin L. and Bromley,  Matthew and Morton,  Charles G. and Minor,  Timothy},
   language = {en},
-  title = {Remote Sensing Estimates of Evapotranspiration from Irrigated Agriculture,  Northwestern Nevada and Northeastern California},
+  title = {{Remote Sensing Estimates of Evapotranspiration from Irrigated Agriculture,  Northwestern Nevada and Northeastern California}},
   publisher = {Desert Research Institute},
   year = {2018},
   month = dec
@@ -97,7 +97,7 @@
   url = {https://rgdoi.net/10.13140/RG.2.2.21605.88808},
   author = {Huntington,  Justin L. and Pearson,  Christopher and Minor,  Blake and Volk,  John and Morton,  Charles G. and Melton,  Forrest and Allen,  Richard G.},
   language = {en},
-  title = {Upper Colorado River Basin OpenET Intercomparison Summary},
+  title = {{Upper Colorado River Basin OpenET Intercomparison Summary}},
   publisher = {Desert Research Institute},
   year = {2022},
   month = jan
@@ -118,7 +118,7 @@
 }
 
 @article{mcevoy2022,
-  title = {Validation and Bias Correction of Forecast Reference Evapotranspiration for Agricultural Applications in Nevada},
+  title = {{Validation and Bias Correction of Forecast Reference Evapotranspiration for Agricultural Applications in Nevada}},
   volume = {148},
   ISSN = {1943-5452},
[…]
