### 2024-05-13 0a2aebb
additional fixes to titles in bibliography
