+++ paper/paper.bib
@@ -204,7 +204,7 @@
 
 
 @book{asce-ewri2005,
-  title = {The ASCE Standardized Reference Evapotranspiration Equation},
+  title = {{The ASCE Standardized Reference Evapotranspiration Equation}},
   ISBN = {9780784475638},
   url = {https://ascelibrary.org/doi/book/10.1061/9780784408056},
   DOI = {10.1061/9780784408056},
@@ -217,7 +217,7 @@
 
 @misc{RefET,
   author = {Morton, Charles G.},
-  title = {ASCE Standardized Reference Evapotranspiration (ET)},
+  title = {{ASCE Standardized Reference Evapotranspiration (ET)}},
   year = {2022},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -232,7 +232,7 @@
 }
 
 @misc{allen2000,
-  title = {REF-ET: Reference evapotranspiration calculation software for FAO and ASCE standardized equations},
+  title = {{REF-ET: Reference Evapotranspiration Calculation Software for FAO and ASCE Standardized Equations}},
   author = {Allen, Richard G.},
   journal = {University of Idaho},
   year = {2000},
