### 2024-05-13 #6368 @cwdunkerly
@kthyng my apologies, I was focusing too much on author names. I have reviewed the full bibliography and the titles are now formatted correctly.

Some sources have titles that capitalize every word and some do not, I have made what is in the bibliography match the source titles.

Please let me know what else I can do.
