### 2024-05-24 #6368 @kthyng
Congratulations on your new publication @cwdunkerly! Many thanks to @cheginit and to reviewers @dostuffthatmatters and @dlebauer for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
